Separate api_isolate_model from data_isolate_model
Ticket: #199
