+++ src/app/Main/Body/QueryPage/Results/Content-Results.component.tsx
@@ -4,7 +4,7 @@ import { FilterContext } from "../../../../Shared/Context/FilterContext";
 import { FilterType } from "../../../../Shared/Model/Filter.model";
 import { TableContext } from "../../../../Shared/Context/TableContext";
 import { DataContext } from "../../../../Shared/Context/DataContext";
-import { DbCollection } from "../../../../Shared/Model/Isolate.model";
+import { DbCollection } from "../../../../Shared/Model/Client_Isolate.model";
 import { filterData } from "../../../../Core/FilterServices/filterData.service";
 import { getValuesOfOneFilterAttribute } from "../../../../Core/FilterServices/getValuesOfOneFilterAttribute.service";
 import { CheckIfFilterIsSet } from "../../../../Core/FilterServices/checkIfFilterIsSet.service";
