+++ src/app/Shared/Model/Api_Isolate.model.ts
@@ -167,13 +167,3 @@ export interface IsolateDTO {
         };
     }[];
 }
-
-export type DbCollection = {
-    microorganism: string;
-    samplingContext: string;
-    matrix: string;
-}[];
-
-export type DbKey = "microorganism" | "samplingContext" | "matrix";
-
-export const DbValues: DbKey[] = ["microorganism", "samplingContext", "matrix"];
