+++ src/app/Shared/Model/Client_Isolate.model.ts
@@ -1,0 +1,9 @@
+export type DbCollection = {
+    microorganism: string;
+    samplingContext: string;
+    matrix: string;
+}[];
+
+export type DbKey = "microorganism" | "samplingContext" | "matrix";
+
+export const DbValues: DbKey[] = ["microorganism", "samplingContext", "matrix"];
